+++ src/server/api/external/openGraph.ts
@@ -13,7 +13,7 @@ export const getOpenGraphMetaImage = async (link_string: string) => {
     //NEXT_PUBLIC_META_API_KEY=http://localhost:3000/api/open-graph 이거 안써두 된다
     const { data: openGraphInfo } = await axios.post(window?.location?.origin + '/api/open-graph', data);
 
-    return openGraphInfo as string | undefined;
+    return openGraphInfo as string;
   } catch (error) {
     // 기존 default image로 대체하기 위해 return null
     return null;
